move inputs into components folder
