+++ config/schema.json
@@ -56,6 +56,10 @@
         "required": {
           "type": "boolean"
         },
+        "preserveDataOnReset": {
+          "type": "boolean",
+          "description": "If true, field data will not be cleared when the scouting form is reset."
+        },
         "code": {
           "type": "string"
         },
