+++ src/components/inputs/CounterInput.tsx
